move export logic to dataset
This cannot work for ds in mem
