+++ qcodes/dataset/data_set.py
@@ -3,15 +3,19 @@ from __future__ import annotations
 import importlib
 import json
 import logging
+import sys
 import time
 import uuid
 from collections.abc import Callable, Mapping, Sequence
 from dataclasses import dataclass
+from pathlib import Path
 from queue import Queue
 from threading import Thread
 from typing import TYPE_CHECKING, Any
 
 import numpy
+from dask.diagnostics.progress import ProgressBar
+from tqdm import tqdm
 
 import qcodes
 from qcodes.dataset.data_set_protocol import (
@@ -1456,6 +1460,45 @@ class DataSet(BaseDataSet):
             add_data_to_dynamic_columns(conn, self.run_id, {tag: data})
 
         self._export_info = export_info
+
+    def _export_as_netcdf(self, path: Path, file_name: str) -> Path:
+        """Export data as netcdf to a given path with file prefix"""
+        import xarray as xr
+
+        if self._estimate_ds_size() > 1000:
+            file_path = path / file_name
+            print("large dataset export.")
+            temp_dir = Path(f"./temp_export_{self.guid}/")
+            temp_dir.mkdir(exist_ok=True)
+            for i in tqdm(range(len(self))):
+                self.to_xarray_dataset(start=i + 1, end=i + 1).to_netcdf(
+                    str(temp_dir / f"ds_{i:03d}.nc"), engine="h5netcdf"
+                )
+            files = [f for f in temp_dir.glob("*.nc")]
+            data = xr.open_mfdataset(files)
+            write_job = data.to_netcdf(file_path, compute=False, engine="h5netcdf")
+            with ProgressBar():
+                print(f"Writing to {file_path}")
+                write_job.compute()
+        else:
+            file_path = super()._export_as_netcdf(path=path, file_name=file_name)
+        return file_path
+
+    def _estimate_ds_size(self) -> float:
+        """
+        Give an estimated size of the dataset as the size of a single row
+        times the len of the dataset. Result is returned in Mega Bytes.
+
+        Note that this does not take overhead into account so it works best
+        if the row size is "large"
+        """
+        sample_data = self.get_parameter_data(start=1, end=1)
+        row_size = 0.0
+
+        for param_data in sample_data.values():
+            for array in param_data.values():
+                row_size += sys.getsizeof(array)
+        return row_size * len(self) / 1024 / 1024
 
     @staticmethod
     def _warn_if_set(
